backend/app/requirements.txt
# AI News Tracker - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0

# AI/ML
openai>=1.3.0
numpy>=1.24.0

# HTTP Requests
httpx>=0.25.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dateutil>=2.8.0

sqlite-vec

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4